requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-html==4.1.1
locust==2.24.0
requests==2.31.0
python-dotenv==1.0.1
webdriver-manager==4.0.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0 